Add app logo and clean up frontend lint
