+++ frontend/src/components/OneSearchLogo.tsx
@@ -1,0 +1,34 @@
+// Copyright (C) 2025 demigodmode
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import { useId, type SVGProps } from 'react'
+
+export function OneSearchLogo(props: SVGProps<SVGSVGElement>) {
+  const clipId = useId()
+  const accentFill = 'hsl(var(--brand-h) var(--brand-s) var(--brand-l))'
+  const darkFill = 'hsl(var(--brand-h) 48% 18%)'
+
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 1500 1499.999933"
+      role="img"
+      aria-label="OneSearch logo"
+      {...props}
+    >
+      <defs>
+        <clipPath id={clipId}>
+          <path d="M 66.339844 67.433594 L 1391 67.433594 L 1391 1390 L 66.339844 1390 Z M 66.339844 67.433594" />
+        </clipPath>
+      </defs>
+      <path fill={accentFill} d="M 1025.253906 597.859375 C 1025.253906 616.0625 1010.472656 630.84375 992.269531 630.84375 L 423.882812 630.84375 C 405.675781 630.84375 390.894531 616.0625 390.894531 597.859375 C 390.894531 579.65625 405.675781 564.875 423.882812 564.875 L 992.269531 564.875 C 1010.472656 564.875 1025.253906 579.65625 1025.253906 597.859375 Z M 1025.253906 597.859375" fillRule="evenodd" />
+      <path fill={accentFill} d="M 1012.566406 467.171875 C 1012.566406 485.378906 997.785156 500.15625 979.578125 500.15625 L 675.085938 500.15625 C 656.878906 500.15625 642.101562 485.378906 642.101562 467.171875 C 642.101562 448.96875 656.878906 434.191406 675.085938 434.191406 L 979.578125 434.191406 C 997.785156 434.191406 1012.566406 448.96875 1012.566406 467.171875 Z M 1012.566406 467.171875" fillRule="evenodd" />
+      <path fill={accentFill} d="M 1025.253906 737.40625 C 1025.253906 755.609375 1010.472656 770.386719 992.269531 770.386719 L 423.882812 770.386719 C 405.675781 770.386719 390.894531 755.609375 390.894531 737.40625 C 390.894531 719.199219 405.675781 704.421875 423.882812 704.421875 L 992.269531 704.421875 C 1010.472656 704.421875 1025.253906 719.199219 1025.253906 737.40625 Z M 1025.253906 737.40625" fillRule="evenodd" />
+      <path fill={accentFill} d="M 1025.253906 881.191406 C 1025.253906 899.398438 1010.472656 914.175781 992.269531 914.175781 L 423.882812 914.175781 C 405.675781 914.175781 390.894531 899.398438 390.894531 881.191406 C 390.894531 862.988281 405.675781 848.210938 423.882812 848.210938 L 992.269531 848.210938 C 1010.472656 848.210938 1025.253906 862.988281 1025.253906 881.191406 Z M 1025.253906 881.191406" fillRule="evenodd" />
+      <g clipPath={`url(#${clipId})`}>
+        <path fill={darkFill} d="M 620.882812 89.625 C 635.109375 106.753906 639.472656 104.734375 640.128906 121.6875 C 640.578125 133.371094 640.390625 132.492188 640.277344 223.976562 C 640.242188 254.171875 652.136719 250.398438 653.324219 252.117188 C 1168.933594 224.722656 1288.019531 946.242188 802.878906 1093.136719 C 576.648438 1161.640625 240.277344 1017.25 255.082031 644.679688 C 252.125 643.621094 258.632812 636.34375 255.679688 635.296875 C 255.257812 635.144531 255.34375 638.894531 120.554688 638.082031 C 106.890625 637.996094 103.21875 634.753906 101.816406 634.78125 C 90.496094 635.003906 91.574219 633.277344 80.292969 631.621094 C 17.625 728.59375 159.335938 1265.261719 643.746094 1292.402344 C 928.644531 1308.367188 1026.121094 1153.976562 1062.214844 1188.957031 C 1235.5 1356.867188 1266.792969 1435.644531 1341.835938 1360.972656 C 1413.550781 1289.609375 1401.214844 1268.5625 1336.601562 1204.132812 C 1184.65625 1052.640625 1176.851562 1046.585938 1191.261719 1024.914062 C 1469.257812 606.894531 1163.730469 64.1875 681.320312 67.449219 C 647.632812 67.679688 646.566406 67.335938 620.882812 89.625 Z M 620.882812 89.625" fillRule="evenodd" />
+      </g>
+      <path fill={accentFill} d="M 620.882812 89.625 C 513.867188 191.445312 119.507812 585.148438 80.292969 631.621094 C 91.574219 633.277344 90.496094 635.003906 101.816406 634.78125 C 103.21875 634.753906 106.890625 637.996094 120.554688 638.082031 C 255.34375 638.894531 255.257812 635.144531 255.679688 635.296875 C 258.632812 636.34375 252.125 643.621094 255.082031 644.679688 C 352.300781 610.726562 619.601562 710.535156 640.4375 504 C 642.15625 486.953125 637.867188 261.40625 644.632812 254.675781 C 646.773438 252.546875 651.183594 254.246094 653.324219 252.117188 C 652.136719 250.398438 640.242188 254.171875 640.277344 223.976562 C 640.390625 132.492188 640.578125 133.371094 640.128906 121.6875 C 639.472656 104.734375 635.109375 106.753906 620.882812 89.625 Z M 620.882812 89.625" fillRule="evenodd" />
+    </svg>
+  )
+}
