+++ frontend/src/contexts/ThemeContextValue.ts
@@ -1,0 +1,14 @@
+// Copyright (C) 2025 demigodmode
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import { createContext } from 'react'
+import type { ThemePreset } from '@/contexts/theme'
+
+export interface ThemeContextType {
+  theme: ThemePreset
+  customHue: number | null  // null = preset active, non-null = slider active
+  setPreset: (preset: ThemePreset) => void
+  setCustomHue: (hue: number) => void
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined)
