+++ frontend/src/contexts/useAuth.ts
@@ -1,0 +1,13 @@
+// Copyright (C) 2025 demigodmode
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import { useContext } from 'react'
+import { AuthContext } from '@/contexts/AuthContextValue'
+
+export function useAuth() {
+  const context = useContext(AuthContext)
+  if (context === undefined) {
+    throw new Error('useAuth must be used within an AuthProvider')
+  }
+  return context
+}
