+++ frontend/src/pages/SearchPage.tsx
@@ -6,7 +6,8 @@ import { useState, useEffect, useRef, useMemo, useCallback } from 'react'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 import { Search, FileText, FileCode, File, Loader2, AlertCircle, ChevronLeft, ChevronRight, Eye } from 'lucide-react'
 import { useSearch, useSources } from '@/hooks/useApi'
-import { useSearchSettings, SNIPPET_LENGTH_MAP } from '@/contexts/SearchSettingsContext'
+import { SNIPPET_LENGTH_MAP } from '@/contexts/searchSettings'
+import { useSearchSettings } from '@/contexts/useSearchSettings'
 import type { SearchResult } from '@/types/api'
 import { cn, sanitizeSnippet, formatSize, formatTimestamp } from '@/lib/utils'
